fix(ctv): overfund factory → dist TX pays its own fee (skip CPFP for v0)
CI revealed two findings on this iteration:

(1) submitpackage rejected our V2 parent with "min relay fee not met,
    0 < 512".  Confirms BIP-431/433 V3/TRUC is required for 0-fee
    parents with ephemeral anchors on Bitcoin Core 28.x.  Adding V3
    to the C library is non-trivial (TH commits to nVersion=2 today).

(2) Sidestep: the BIP-119 TH commits to dist TX OUTPUTS, not its input
    amount.  If the LSP overfunds the factory by an extra `--dist-tx-
    fee-sats` margin (default 4000 sat), the dist TX gets an implicit
    fee equal to (funding_input − sum_committed_outputs).  We can
    sendrawtransaction it directly — no CPFP child, no submitpackage,
    no V3 needed.  Production deployments still want V3+CPFP for fee
    flexibility, but for a v0 demo this is the right call.

Other changes:
  - cppcheck: cast CTV_FACTORY_MAX_USERS_SINGLE_LAYER to (unsigned int)
    in two %u format strings (was implicit signed-int → invalid).
  - Renamed --package-fee-sats → --dist-tx-fee-sats; harness updated.
  - Step counters 1/8 → 1/7 reflecting the simpler flow.
  - build_cpfp_child kept in the source but no longer invoked; future
    V3/CPFP work resurrects it.

+++ tools/ctv_factory_build.c
@@ -61,7 +61,7 @@ static void usage(const char *prog) {
         "  --funding-txid T         32-byte funding outpoint (hex, LE wire order)\n"
         "  --funding-vout V         funding outpoint index\n"
         "    If both txid and vout are set, also emits dist_tx_segwit_hex.\n",
-        prog, CTV_FACTORY_MAX_USERS_SINGLE_LAYER);
+        prog, (unsigned int)CTV_FACTORY_MAX_USERS_SINGLE_LAYER);
 }
 
 int main(int argc, char **argv) {
@@ -100,7 +100,7 @@ int main(int argc, char **argv) {
     }
 
     if (n_users == 0u || n_users > CTV_FACTORY_MAX_USERS_SINGLE_LAYER) {
-        fprintf(stderr, "--users must be 1..%u\n", CTV_FACTORY_MAX_USERS_SINGLE_LAYER);
+        fprintf(stderr, "--users must be 1..%u\n", (unsigned int)CTV_FACTORY_MAX_USERS_SINGLE_LAYER);
         return 2;
     }
     if (slot_deposit == 0u) {
